# server/models/painting/painting.js
import mongoose from 'mongoose';
import normalize from 'normalize-mongoose';

const Schema = mongoose.Schema;

const paintingSchema = new Schema({
    data: Buffer,
    name: String,
    contentType: String
}, {collection: 'painting'});

paintingSchema.plugin(normalize);

const Painting = mongoose.model('Painting', paintingSchema);

export default Painting;