# server/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "Bohian-art API"
host: "localhost:3000"
basePath: "/"
tags:
  - name: "Users"
    description: "User management"
  - name: "Paintings"
    description: "Painting Management"
schemes:
  - "https"
  - "http"
paths:
  /users/auth:
    post:
      summary: "Authorize user"
      operationId: "auth"
      tags:
        - "Users"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/AuthUserResponse"
        401:
          description: "Invalid credentials"
  /users:
    get:
      summary: "Get all users"
      operationId: "getAllUsers"
      tags:
        - "Users"
      security:
        - bearer_auth: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/UsersResponse"
        401:
          description: "Unauthorized"
    post:
      summary: "Create a new user"
      operationId: "createUser"
      tags:
        - "Users"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/UserResponse"
        400:
          description: "Such user have already created"
  /users/{userId}:
    get:
      summary: "Get the user by id"
      operationId: "getUser"
      tags:
        - "Users"
      security:
        - bearer_auth: []
      parameters:
        - name: "userId"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/UserResponse"
        401:
          description: "Unauthorized"
    put:
      summary: "Update the user"
      operationId: "updateUser"
      tags:
        - "Users"
      security:
        - bearer_auth: []
      parameters:
        - name: "userId"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/UserResponse"
        401:
          description: "Unauthorized"
    delete:
      summary: "Delete the user"
      operationId: "deleteUser"
      tags:
        - "Users"
      security:
        - bearer_auth: []
      parameters:
        - name: "userId"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/BaseResponse"
        401:
          description: "Unauthorized"
  /paintings:
    get:
      summary: "Get all paintings"
      operationId: "getAllPaintings"
      tags:
        - "Paintings"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/PaintingsResponse"
        401:
          description: "Unauthorized"
    post:
      summary: "Upload a new painting"
      operationId: "uploadPainting"
      tags:
        - "Paintings"
      security:
        - bearer_auth: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/Painting"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/PaintingResponse"
        401:
          description: "Unauthorized"
  /paintings/{paintingId}:
    get:
      summary: "Get the painting by id"
      operationId: "getPainting"
      tags:
        - "Paintings"
      parameters:
        - name: "paintingId"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/PaintingResponse"
        401:
          description: "Unauthorized"
    delete:
      summary: "Delete the painting"
      operationId: "deletePainting"
      tags:
        - "Paintings"
      security:
        - bearer_auth: []
      parameters:
        - name: "paintingId"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/BaseResponse"
        401:
          description: "Unauthorized"
securityDefinitions:
  bearer_auth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  BaseResponse:
    type: "object"
    properties:
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "success"
      - "errorMessage"
  AuthUserResponse:
    type: "object"
    properties:
      user:
        $ref: "#/definitions/User"
      token:
        type: "string"
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "user"
      - "token"
      - "success"
      - "errorMessage"
  User:
    type: "object"
    properties:
      id:
        type: "string"
      username:
        type: "string"
      password:
        type: "string"
        format: "password"
    required:
      - "username"
      - "password"
  UserResponse:
    type: "object"
    properties:
      user:
        $ref: "#/definitions/User"
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "user"
      - "success"
      - "errorMessage"
  UsersResponse:
    type: "object"
    properties:
      users:
        type: "array"
        items:
          $ref: "#/definitions/User"
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "users"
      - "success"
      - "errorMessage"
  Painting:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      contentType:
        type: "string"
      data:
        type: "object"
    required:
      - "name"
      - "contentType"
      - "data"
  PaintingResponse:
    type: "object"
    properties:
      painting:
        $ref: "#/definitions/Painting"
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "painting"
      - "success"
      - "errorMessage"
  PaintingsResponse:
    type: "object"
    properties:
      paintings:
        type: "array"
        items:
          $ref: "#/definitions/Painting"
      success:
        type: "boolean"
      errorMessage:
        type: "string"
    required:
      - "users"
      - "success"
      - "errorMessage"